+++ pypeit/arcimage.py
@@ -11,13 +11,13 @@ import numpy as np
 from pypeit import msgs
 from pypeit import masterframe
 from pypeit.par import pypeitpar
-from pypeit.images import combinedimage
+from pypeit.images import calibrationimage
 from pypeit.core import procimg
 
 from IPython import embed
 
 
-class ArcImage(combinedimage.CombinedImage, masterframe.MasterFrame):
+class ArcImage(calibrationimage.CalibrationImage, masterframe.MasterFrame):
     """
     Generate an Arc Image by processing and combining one or more arc frames.
 
@@ -59,37 +59,19 @@ class ArcImage(combinedimage.CombinedImage, masterframe.MasterFrame):
         self.par = pypeitpar.FrameGroupPar(self.frametype) if par is None else par
 
         # Start us up
-        combinedimage.CombinedImage.__init__(self, spectrograph, det, self.par['process'],
-                                             files=files, frametype=self.frametype)
+        calibrationimage.CalibrationImage.__init__(self, spectrograph, det, self.par['process'],
+                                                   files=files, frametype=self.frametype)
 
         # MasterFrames: Specifically pass the ProcessImages-constructed
         # spectrograph even though it really only needs the string name
         masterframe.MasterFrame.__init__(self, self.master_type, master_dir=master_dir,
                                          master_key=master_key, reuse_masters=reuse_masters)
 
-    def build_image(self):
-        """
-        Build the arc image from one or more arc files.
-
-        If we ever decide to apply the gain, we will need
-        to worry about saturation
-
-        Args:
-
-        Returns:
-            `numpy.ndarray`_: Combined, processed image
-            
-        """
-        # Load
-        self.load_images()
         # Process steps
-        process_steps = procimg.init_process_steps(self.msbias, self.par['process'])
-        process_steps += ['trim']
+        self.process_steps = procimg.init_process_steps(self.msbias, self.par['process'])
+        self.process_steps += ['trim']
+        self.process_steps += ['orient']
         # NOT applying gain to deal 'properly' with saturation
-        # Process
-        self.process_images(process_steps, bias=self.msbias)
-        # Combine + Return
-        return self.combine()
 
     def save(self, outfile=None, overwrite=True):
         """
@@ -103,11 +85,11 @@ class ArcImage(combinedimage.CombinedImage, masterframe.MasterFrame):
                 Overwrite any existing file.
         """
         super(ArcImage, self).save(self.image, 'ARC', outfile=outfile, overwrite=overwrite,
-                                   raw_files=self.file_list, steps=self.steps)
+                                   raw_files=self.file_list, steps=self.process_steps)
 
     # TODO: it would be better to have this instantiate the full class
     # as a classmethod.
-    def load_arcimage(self, ifile=None, return_header=False):
+    def load(self, ifile=None, return_header=False):
         """
         Load the arc frame data from a saved master frame.
 
@@ -122,5 +104,5 @@ class ArcImage(combinedimage.CombinedImage, masterframe.MasterFrame):
             Returns a `numpy.ndarray`_ with the arc master frame image.
             Also returns the primary header, if requested.
         """
-        return super(ArcImage, self).load_master('ARC', ifile=ifile, return_header=return_header)
+        return super(ArcImage, self).load('ARC', ifile=ifile, return_header=return_header)
