+++ pypeit/masterframe.py
@@ -163,7 +163,7 @@ class MasterFrame(object):
     # TODO: Add a base-level staticmethod one-liner?
     # TODO: include checksum keyword, used to validate data when
     # loading?
-    def load_master(self, ext, ifile=None, return_header=False):
+    def load(self, ext, ifile=None, return_header=False):
         """
         Generic master file reader.
