+++ pypeit/scienceimages.py
