+++ pypeit/traceimage.py
@@ -5,11 +5,11 @@ import numpy as np
 
 from pypeit import msgs
 from pypeit.par import pypeitpar
-from pypeit.images import combinedimage
+from pypeit.images import calibrationimage
 from pypeit.core import procimg
 
 
-class TraceImage(combinedimage.CombinedImage):
+class TraceImage(calibrationimage.CalibrationImage):
     """
     Generate the image for tracing the slits/orders, typically from a
     set of flat-field or twilight sky exposures
@@ -28,18 +28,14 @@ class TraceImage(combinedimage.CombinedImage):
     # Frametype is a class attribute
     frametype = 'trace_image'
 
-    def __init__(self, spectrograph, files=None, det=1, par=None):
+    def __init__(self, spectrograph, files=None, det=1, par=None, bias=None):
         self.par = pypeitpar.FrameGroupPar('trace') if par is None else par
         # Start us up
-        combinedimage.CombinedImage.__init__(self, spectrograph, det, self.par['process'],
-                                             files=files, frametype=self.frametype)
+        calibrationimage.CalibrationImage.__init__(self, spectrograph, det, self.par['process'],
+                                                   files=files, frametype=self.frametype)
+        # Processing steps
+        self.process_steps = procimg.init_process_steps(bias, self.par['process'])
+        self.process_steps += ['trim']
+        self.process_steps += ['apply_gain']
+        self.process_steps += ['orient']
 
-    def build_image(self, bias=None):
-        # Load
-        self.load_images()
-        # Process + combine
-        process_steps = procimg.init_process_steps(bias, self.proc_par)
-        process_steps += ['trim']
-        process_steps += ['apply_gain']
-        self.process_images(process_steps, bias=bias)
-        self.image = self.combine()
